wip: cleanup app/appv2 control signal fot loginSUpported api
