+++ jest.config.ts
@@ -42,7 +42,7 @@ const config: Config = {
 	coverageProvider: 'babel',
 
 	// A list of reporter names that Jest uses when writing coverage reports
-	coverageReporters: ['text', 'cobertura', 'lcov'],
+	coverageReporters: ['lcov', 'html'],
 
 	// An object that configures minimum threshold enforcement for coverage results
 	// coverageThreshold: {
